+++ libs/@local/harpc/client/typescript/src/net/Transport.ts
@@ -1,19 +1,10 @@
 import type { DNS } from "@multiformats/dns";
-import { Data } from "effect";
 
 import type { NoiseConfig, TCPConfig, YamuxConfig } from "./Config.js";
 import type * as internal from "./internal/transport.js";
 
-export { TransportError } from "./internal/transport.js";
+export { InitializationError, TransportError } from "./internal/transport.js";
 export { type Multiaddr, multiaddr } from "@multiformats/multiaddr";
-
-export class InitializationError extends Data.TaggedError(
-  "InitializationError",
-)<{ cause: unknown }> {
-  get message() {
-    return "Failed to initialize client";
-  }
-}
 
 export type Address = internal.Address;
